genealogia_saas/backend/requirements.txt
# Dependências do Backend

# Framework Web
fastapi>=0.68.0
uvicorn>=0.15.0

# Banco de Dados
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.1

# Autenticação e Segurança
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Validação e Serialização
pydantic>=1.8.2
email-validator>=1.1.3

# Utilitários
python-dotenv>=0.19.0
tenacity>=8.0.1

# Testes
pytest>=6.2.5
httpx>=0.19.0